Revert "initialize setting"
This reverts commit 4b99f483ab50aadcb64045a65246b61cdc9e4cfa.

+++ Project/Engine/structs.h
@@ -29,7 +29,7 @@ struct tMtrlData
 	Vec4 v4Arr[4] = {};
 	Matrix matArr[4] = {};
 
-	int	bTex[(UINT)TEX_PARAM::END];
+	int	bTex[TEX_PARAM::END];
 
 	int iPadding[2];
 };
